requirements.txt
flask
requests
scikit-learn
pytesseract
opencv-python
Pillow
pyzbar
tldextract
python-whois